refactor(painters/dom): unify drawing block rendering across renderer and table cells
Move renderDrawingContent (vectorShape, chart, shapeGroup, shapeText
paths) out of renderer.ts into drawings/renderDrawingContent.ts so the
main renderer and table cells share a single rendering implementation.
Extract the table-cell drawing wrapper into drawings/tableDrawingFrame.ts
to deduplicate the in-flow and anchored drawing scaffolding, and thread
the renderDrawingContent callback through renderTableCell so non-image
drawings render via the shared path. Add unit coverage for the new
renderDrawingContent module and the table cell frame builder.
